+++ packages/sling-web-component-form/src/state/constant.js
@@ -1,0 +1,1 @@
+export const FORM = '__FORM__';
